# pendataan-pasien-covid/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Patient Corona  Virus",
        "description": "API Patient Corona Virus",
        "contact": {
            "name": "Rifky",
            "email": "rifky@gmail.com"
        },
        "license": {
            "name": "Your License",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Get list of patients",
                "description": "Returns a list of patients with optional filtering and pagination.",
                "operationId": "getPatients",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter by patient name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[address]",
                        "in": "query",
                        "description": "Filter by patient address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Filter by patient status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[sort]",
                        "in": "query",
                        "description": "Sort the result by a specific field",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[order]",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page[limit]",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Create a new patient",
                "description": "Create a new patient with the provided information",
                "operationId": "storePatient",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "positif/negatif/meninggal"
                                    },
                                    "in_date_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "out_date_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/patients/{id}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient by ID",
                "description": "Update an existing patient with the provided information",
                "operationId": "showDataPatient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the patient to be Show Data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully Get Data Patient"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient by ID",
                "description": "Update an existing patient with the provided information",
                "operationId": "updatePatient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the patient to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "positif/negatif/meninggal"
                                    },
                                    "in_date_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "out_date_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated successfully"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient by ID",
                "description": "Update an existing patient with the provided information",
                "operationId": "DeletePatient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the patient to be Deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully Get Data Patient"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}